Add files via upload
PREDICTION USING SUPERVISED ML
Predict the percentage of an student based on the number of study hours.

Simple Linear Regression : In this regression task we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied. This is a simple linear regression task as it involves just two variables.