Add files via upload
Problem Statement : We are trying to predict the stock price of a particular company by taking the company's historical stock price data(yahoo finance) and try to see if the historical data has any impact. We also tried to anlyze the news headlines of the previous day, gets its sentiment and try to combine the information that we got with the historical data and build a hyrbid model that predicts Stock Price.

Implementation: for Time series approach, we used the ARIMA model, which is a powerful model for forecasting future values of a time series based on its past values. For the machine learning approach, we tried out various ML Algorithms.(such as linear regression,mean squared error, Decission tree clssifier, random forest,XGB regressor& AdaBoostRegressor etc). we also incorporated textual analysis into my model by extracting sentiment scores from news headlines on a particular day. found that these sentiment scores improved the model's accuracy by a greater extent.(NLTK, sentiment intensity analyser,min max scaler and text blob etc).

Note/ Termenolgy: Notes: 

1.Statsmodels is a Python package that allows users to explore data, estimate statistical models, and perform statistical tests. An extensive list of descriptive statistics, statistical tests, plotting functions, and result statistics are available for different types of data and each estimator. It complements SciPy's stats module.

2.AdaBoost, short for Adaptive Boosting, is a statistical classification meta-algorithm.
Learn how to use an AdaBoost regressor to improve the accuracy of a prediction based on a set of data. This class implements the AdaBoost.R2 algorithm and provides parameters, methods, and references for this class.

3.XGBoost is an open-source software library which provides a regularizing gradient boosting framework,The XGBoost library was developed to provide high computing power during training and to parallelize and accelerate the training.

4.Lemmatization in linguistics is the process of grouping together the inflected forms of a word so they can be analysed as a single item, identified by the word's lemma, or dictionary form.
Unlike stemming, lemmatization uses the part of speech tags and the meaning of the word in the sentence to see the main context of the document.

5.TextBlob is a python Library which stands on the NLTK .It works as a framework for almost all necessary task , we need in Basic NLP ( Natural Language Processing ) 

6.Stop words are the words in a stop list which are filtered out before or after processing of natural language data because they are insignificant.

7.SentimentIntensityAnalyzer is a tool that can measure the sentiment of a text, such as positive, negative, or neutral. It is based on a lexicon of words that have different scores for how they express emotion. It also uses some rules to adjust the scores based on the context of the text. For example, it can detect negations, intensifiers, and modifiers that can change the sentiment of a word or a phrase.
SentimentIntensityAnalyzer is part of the Natural Language Toolkit (NLTK) library for Python, which is a collection of modules and resources for natural language processing and analysis.

8.MinMaxScaler from the scikit-learn library:
MinMaxScaler is a class in the sklearn.preprocessing module that scales and translates each feature individually such that it is in the given range on the training set, e.g. between zero and one.

9.seasonal_decompose:A useful abstraction for selecting forecasting methods is to break a time series down into systematic and unsystematic components.

10.ARIMA- Autoregressive integrated moving average- In statistics and econometrics, and in particular in time series analysis, an autoregressive integrated moving average model is a generalization of an autoregressive moving average model.A time series is a sequence where a metric is recorded over regular time intervals. Depending on the frequency, a time series can be of yearly.

11.SARIMAX is a type of time series model that can capture both seasonal and non-seasonal patterns in the data. SARIMAX stands for Seasonal Auto-Regressive Integrated Moving Average with eXogenous factors. It is an extension of the ARIMA model, which is a popular model for forecasting univariate time series. SARIMAX can also incorporate external variables that may affect the time series, such as weather, holidays, or economic indicators.


conclusion: #
1.From the above results, the linear regression still outperformes all other tree based and ensemble methods. As we have seen in the top the stock price has a strong relationship to its immediate lag stock price and thus it is able to capture/ predict the price accurately with its historical data

2.Adding the sentiment analysis of the news headlines even though it wasn't that relevant seemed to improve the

  
 accuracy in the Random Forest Regressor when we compared it to the above model where we trained it without the news headlines data. But it is important to note that we used different ranges (sizes) of training and test data in both the case.